Outliers: * Move ListTypes from statistics package to listtypes package
