+++ Numbers/src/main/kotlin/mathtools/numbers/statistics/Statistics.kt
@@ -1,5 +1,6 @@
 package mathtools.numbers.statistics
 
+import mathtools.numbers.listtypes.listSum
 import kotlin.math.sqrt
 
 /** Key Statistics Functions */
